# Speedup two tests