--- a/nltk/test/gensim.doctest
+++ b/nltk/test/gensim.doctest
@@ -22,7 +22,8 @@ Train the model
 Here we train a word embedding using the Brown Corpus:
 
     >>> from nltk.corpus import brown
-    >>> model = gensim.models.Word2Vec(brown.sents())
+    >>> train_set = brown.sents()[:10_000]
+    >>> model = gensim.models.Word2Vec(train_set)
 
 It might take some time to train the model. So, after it is trained, it can be saved as follows: