--- a/nltk/test/unit/test_rte_classify.py
+++ b/nltk/test/unit/test_rte_classify.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-import unittest
+import pytest
 
 from nltk.corpus import rte as rte_corpus
 from nltk.classify.rte_classify import RTEFeatureExtractor, rte_features, rte_classifier
@@ -56,7 +56,7 @@ word_overlap    => 1
 """
 
 
-class RTEClassifierTest(unittest.TestCase):
+class TestRTEClassifier:
     # Test the feature extraction method.
     def test_rte_feature_extraction(self):
         pairs = rte_corpus.pairs(['rte1_dev.xml'])[:6]
@@ -68,24 +68,27 @@ class RTEClassifierTest(unittest.TestCase):
         expected_output = expected_from_rte_feature_extration.strip().split('\n')
         # Remove null strings.
         expected_output = list(filter(None, expected_output))
-        self.assertEqual(test_output, expected_output)
+        assert test_output == expected_output
 
     # Test the RTEFeatureExtractor object.
     def test_feature_extractor_object(self):
         rtepair = rte_corpus.pairs(['rte3_dev.xml'])[33]
         extractor = RTEFeatureExtractor(rtepair)
-        self.assertEqual(extractor.hyp_words, {'member', 'China', 'SCO.'})
-        self.assertEqual(extractor.overlap('word'), set())
-        self.assertEqual(extractor.overlap('ne'), {'China'})
-        self.assertEqual(extractor.hyp_extra('word'), {'member'})
+
+        assert extractor.hyp_words == {'member', 'China', 'SCO.'}
+        assert extractor.overlap('word') == set()
+        assert extractor.overlap('ne') == {'China'}
+        assert extractor.hyp_extra('word') == {'member'}
 
     # Test the RTE classifier training.
     def test_rte_classification_without_megam(self):
-        clf = rte_classifier('IIS')
-        clf = rte_classifier('GIS')
+        # Use a sample size for unit testing, since we
+        # don't need to fully train these classifiers
+        clf = rte_classifier('IIS', sample_N=100)
+        clf = rte_classifier('GIS', sample_N=100)
 
-    @unittest.skip("Skipping tests with dependencies on MEGAM")
+    @pytest.mark.skip("Skipping tests with dependencies on MEGAM")
     def test_rte_classification_with_megam(self):
         nltk.config_megam('/usr/local/bin/megam')
-        clf = rte_classifier('megam')
-        clf = rte_classifier('BFGS')
+        clf = rte_classifier('megam', sample_N=100)
+        clf = rte_classifier('BFGS', sample_N=100)