Philosophically, I don't think unit tests should be responsible for loading whole corpora. Ideally, they should load mock datasets. But for now we can downsample their input size to make them a bit faster.

`rte_classify`: ~60s to 4s
`gensim.doctest`: 45s to 25s

This is not to say this diff shaves 80 seconds off of the whole test suite. It's more around the lines of 30 or 40 seconds.